Renaming fabfile
Rename fabfile to fab_tasks and import this in the fabfile. Makes
it easier to import in other projects.
