+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
  * Move timeout logic from fabric file into utility decorator.
  * Add wait_for_ssh function to the bootstrap commands. This ensures ssh is up before we bootstrap.
+ * Rename fabfile to fab_tasks to make it easier to import in other projects.
 
 ## Version 0.1
